define _Unwind_DeleteException in unwind-avr-sup.c
ok so as it turns out before because I defined DeleteException as inline,
which made it work before. Now I'm defining it as a seperate function.
Doing it this way increases binary bloat, but better conforms to the
itanium abi. Later I'll figure out a way to reduce the exception handling
ABI, but for now just put it in unwind-avr-sup.c