Finally got this to link a program. Most things are stubs right now.
Also, compiling comes with the caveat that it only works if libitm
is disabled and --disable-shared is used. libitm does something
weird in one of its function prologues which fae doesn't cope well
with, and currently I am having difficulty getting the shared library
to work. Also, most functions are stubs that need to actually be implemented.

Also, _Unwind_DeleteException is stubbed, but it should really be using
the unwind-dw2.c implementation. I need to get a autoconf flag to do that
eventually.

Also, libstdc++ built object code doesn't seem to work. It mentions
symbols of fae.idx being referenced in discarded sections of
monetary_members and numpunct when iostream is used, which is weird
considering it has no issue discarding other unused fae sections.
For now I won't investigate since it isn't used in embedded (the primary goal)
anyways, but it's worth thinking about.