Finally got it to unwind functions!
There were many minor changes in this one. First off, the unwind
routines were changed to be positively indexed instead of negatively
for ease of debugging.

Secondly, I discovered it is possible for the compiler to save
caller saved registers, for some reason. This means it allocates
stack space without saving a register, which I didn't account for.

Third, I altered the calling convention of __fae_get_ptr to be
memory based because x86_64 calling convention didn't have enough
register space. This could be altered on other platforms, but for
now this is ok.

Fourth, I found out that the x86 push instruction can be 5 or 4
bytes long, depending on the register. This is kinda dumb, but
that's x86. I padded it with nop to fix the alignment.