ok so actually x86 is stupid, so I'm just giving up on memory calls for now.
In x86, calls where arguments are spilled to registers may generate
pops and pushes in the middle of the function block. More annoyingly,
this means stack usage can vary depending on when in the funciton
you are, which utterly ruins my current stack-unwinding scheme.

Apparently ARM code generation avoids this by forcing it to stack
allocate in the beginning of the function, and using relative
addressing to set those arguments. I half-suspect this could be EABI
mandated, but have no idea. In any case, to fix this I'd have to
change x86 code gen, which is annoying. For now I'm skipping trying
to fix this as memory-spilled arguments are relatively rare. ARM
anyways doesn't have to deal with this, and x86 is mostly just a
demonstration.